Disk read write included

+++ src/driver/vga/vga_term.c
@@ -49,6 +49,7 @@ void vga_init(){
 }
 
 
+
 void clear_screen(){
     size_t pitch_pixels = fb_pitch / sizeof(uint32_t);
     for(size_t row = 0; row < (size_t) (fb_height - 1); row++){
@@ -187,8 +188,8 @@ void backspace_manage() {
     }
 
     // Erase the previous character by drawing a blank space
-    for (int row = 0; row < font_height; row++) {
-        for (int col = 0; col < font_width; col++) {
+    for (uint64_t row = 0; row < font_height; row++) {
+        for (uint64_t col = 0; col < font_width; col++) {
             set_pixel(cur_x + col, cur_y + row, back_color); // Black color (background)
         }
     }
